// frontend/src/lib/api/models/adminChannelCreateRequestContentPrivacy.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

export type AdminChannelCreateRequestContentPrivacy =
  (typeof AdminChannelCreateRequestContentPrivacy)[keyof typeof AdminChannelCreateRequestContentPrivacy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminChannelCreateRequestContentPrivacy = {
  PUBLIC: "PUBLIC",
  UNLISTED: "UNLISTED",
  PRIVATE: "PRIVATE",
} as const;